app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Mactaquac web app"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.11",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
]
